+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file. From versio
 ### Fixed
 
 - Shutdown should wait for in flight requests by @mkleczek in #4702
+- Fix login with uppercase and mixed case role names by @taimoorzaeem in #4678
 
 ### Changed